third_party/blink/web_tests/external/wpt/pointerevents/pointerevent_after_target_removed_mouse-expected.txt

This is a testharness.js-based test.
[FAIL] pointer events from mouse received before/after child removal at pointerdown
  assert_equals: events received expected "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,(child-removed),pointerup@parent,pointerover@parent,pointerdown@parent,pointerup@parent,pointerout@parent,pointerleave@parent" but got "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,(child-removed),pointerover@parent,pointerup@parent,pointerdown@parent,pointerup@parent,pointerout@parent,pointerleave@parent"
Harness: the test ran to completion.
